Connected components to store to receive data. Moved initial data fetch to dashboard. Issue: must refresh screen once after login to prevent Error 401 errors.
