+++ src/components/Summary.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import { useDispatch } from 'react-redux';
-import authAxios from '../utils/authaxios';
+import { useDispatch, connect } from 'react-redux';
+// import authAxios from '../utils/authaxios';
 
 import Table from '@material-ui/core/Table';
 import TableBody from '@material-ui/core/TableBody';
@@ -11,14 +11,14 @@ import Paper from '@material-ui/core/Paper';
 
 
 // Material UI Table --------------------------------------------------
-interface Props {
-  rows: [{
-    name: '',
-    village: '',
-    loan_amt: '',
-    loan_init: ''
-  }]
-};
+// interface Props {
+//   rows: [{
+//     name: '',
+//     village: '',
+//     loan_amt: '',
+//     loan_init: ''
+//   }]
+// };
 
 
 function SimpleTable(props) {
@@ -50,29 +50,39 @@ function SimpleTable(props) {
 }
 //Material UI Table ^^^-------------------------------------------------
 
-const Summary = () => {
-  const dispatch = useDispatch();
- 
-  // TODO: Remove before flight. Also the button that calls it.
+const Summary = props => {
+  // const dispatch = useDispatch();
+  console.log("Summary props:", props);
+  
+  // FIXME: Infinite loop. Use a useEffect?
+  // useEffect(() => {
+  //   authAxios.get('https://tiemendo.herokuapp.com/api/auth/loans')
+  //   .then (res => {
+  //     console.log(res);
+  //     dispatch({
+  //       type: 'LOANS_NEW_LIST',
+  //       payload: res.data
+  //     })
+  //   })
+  //   .catch(err => console.log('Loan list error:', err)
+  //   );
+  // },[dispatch]);
 
-  authAxios.get('https://tiemendo.herokuapp.com/api/auth/loans')
-  .then (res => {
-    console.log(res);
-    dispatch({
-      type: 'LOANS_NEW_LIST',
-      payload: res.data
-    })
-  })
-  .catch(err => console.log('Loan list error:', err)
-  );
 
   return (
     <div className="workspace">
-      <h2>The summary page. Let's summarize.</h2>
+      <h2>Reports of loan information</h2>
 
       {SimpleTable(/* Get the state data here */)}
     </div>
   )
 }
 
-export default Summary;
+const mapStateToProps = (state) => {
+  return {
+    loans: state.loans,
+    clients: state.clients
+  }
+}
+
+export default connect(mapStateToProps)(Summary);
